[feature] update dest to take 'none' as an option, move template object work into templater
